app-engine-microservices/gamma-service/requirements.txt
django-cors-headers==3.4.0
djangorestframework==3.11.0
google-cloud-pubsub==2.2.0
